Deepseek model DOC recog/requirements.txt
streamlit
azure-identity
azure-ai-inference
PyPDF2
python-docx
pandas
pypandoc
deep-translator